refactor insert
